# swagger.yaml
openapi: 3.0.3
info:
  title: Sky's UPC Database
  description: This is a free and open-source UPC database.
  contact:
    email: upc@skystuff.cc
  version: 1.0.0
servers:
  - url: https://upc.skystuff.cc/api/
tags:
  - name: UPC
    description: Everything about your Pets
paths:
  /upc/{upc}:
    get:
      tags:
        - UPC
      summary: Get information about a UPC code.
      parameters:
        - in: path
          name: upc
          schema:
            type: integer
          required: true
          description: 8 or 12 digit UPC code to look up.
      description: Get information about a UPC code.
      operationId: getUPC
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/upc'
        '400':
          description: Invalid UPC
        '404':
          description: UPC Not Found
  /upc/:
    post:
      tags:
        - UPC
      summary: Upload data about a UPC to the database.
      description: Upload data about a UPC to the database.
      operationId: addUPC
      requestBody:
        description: Upload data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/upc'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
  /upc/bulk/:
    get:
      tags:
        - UPC
      summary: Get bulk data for up to 50 UPCs
      description: Returns bulk data for 50 UPCs, and processes them to ensure they are good. You can pass 50 real UPCs, and some invalid ones, and it will filter out the invalid ones.
      operationId: upc.bulk
      parameters:
        - in: query
          name: upcs
          schema:
            type: string
          required: true
          description: Comma separated UPCs
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkUPC'
        '400':
          description: Too many UPCs requested.

  /upc/validate/{upc}:
    get:
      tags:
        - UPC
      summary: Test if a UPC is valid
      description: Returns a boolean on if a UPC is valid or not.
      operationId: upc.validate
components:
  schemas:
    upc:
      type: object
      properties:
        upc:
          type: string
        name:
          type: string
        quantity:
          type: string
        quantity_unit:
          type: string
      required:
        - upc
        - name